refactor: split the list invoices page into multiple components
